requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy[asyncio]
asyncmy
